[akka-dynamodb] Extract WorkerIdAssignItem and be able to generate the table creation request needed for the assignment
